Moved SAM selection classes to sub-package sam
