+++ java/component/keyple-core/src/main/java/org/eclipse/keyple/transaction/MatchingSe.java
@@ -26,7 +26,7 @@ public class MatchingSe {
      * Constructor taking a SeSelector as an argument. Keeps the isKeepChannelOpen flag and the
      * extraInfo for later usage.
      * 
-     * @param seSelectionRequest the se selection request
+     * @param seSelectionRequest the SE selection request
      */
     public MatchingSe(SeSelectionRequest seSelectionRequest) {
         this.channelIsKeptOpen = seSelectionRequest.getSelectionRequest().isKeepChannelOpen();
