+++ java/example/calypso/pc/src/main/java/org/eclipse/keyple/example/calypso/pc/transaction/CalypsoUtilities.java
@@ -17,10 +17,10 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.EnumMap;
 import java.util.Properties;
-import org.eclipse.keyple.calypso.transaction.CalypsoSam;
 import org.eclipse.keyple.calypso.transaction.PoTransaction;
-import org.eclipse.keyple.calypso.transaction.SamSelectionRequest;
-import org.eclipse.keyple.calypso.transaction.SamSelector;
+import org.eclipse.keyple.calypso.transaction.sam.CalypsoSam;
+import org.eclipse.keyple.calypso.transaction.sam.SamSelectionRequest;
+import org.eclipse.keyple.calypso.transaction.sam.SamSelector;
 import org.eclipse.keyple.example.generic.pc.ReaderUtilities;
 import org.eclipse.keyple.seproxy.ChannelState;
 import org.eclipse.keyple.seproxy.SeProxyService;
